+++ service-implementation/src/main/webapp/ui/js/controllers/processor/processorcontroller.js
@@ -1217,8 +1217,20 @@ var rest = myApp.controller(
                       $scope.processorCredProperties = data.processorDefinition.credentialProperties;
                       });
                       $scope.supportedJavaScriptCheckBox = $scope.javaScriptCheckBoxConditionalSweeper.supportedJavaScriptCheckBox;
-                      break;	
-                  case "HTTPSYNCPROCESSOR": 
+                      break;
+				  case "HTTPSYNCPROCESSOR":
+				  	$scope.isProcessorTypeSweeper = false;
+				  	$scope.isProcessorTypeConditionalSweeper = false;
+				  	$scope.isProcessorTypeHTTPListener = true;
+				  	$scope.isProcessorTypeFileWriter = false;
+				  	$scope.isProcessorTypeDropbox = false;
+				  	$scope.processor.protocol = $scope.processorData.supportedProcessors.options[getIndexOfValue($scope.processorData.supportedProcessors.options, $scope.selectedProcessorType)];
+				  	$rootScope.restService.get('data/processor/properties/httpSync.json', function (data) {
+				  	    $scope.separateProperties(data.processorDefinition.staticProperties);
+				  	    $scope.processorCredProperties = data.processorDefinition.credentialProperties;				  	      
+				  	});
+				  	$scope.supportedJavaScriptCheckBox = $scope.javaScriptCheckBox.supportedJavaScriptCheckBox;
+				  	break;
 				  case "HTTPASYNCPROCESSOR": 				 
 					$scope.isProcessorTypeSweeper = false;
 					$scope.isProcessorTypeConditionalSweeper = false;
@@ -1226,7 +1238,7 @@ var rest = myApp.controller(
 					$scope.isProcessorTypeFileWriter = false;
 					$scope.isProcessorTypeDropbox = false;
 					$scope.processor.protocol = $scope.processorData.supportedProcessors.options[getIndexOfValue($scope.processorData.supportedProcessors.options, $scope.selectedProcessorType)];
-					$rootScope.restService.get('data/processor/properties/httpsyncAndAsync.json', function (data) {						
+					$rootScope.restService.get('data/processor/properties/httpAsync.json', function (data) {						
 					  $scope.separateProperties(data.processorDefinition.staticProperties);
 					  $scope.processorCredProperties = data.processorDefinition.credentialProperties;
 					});